[fix]회원가입 css 수정 및 디렉토리 구조 변경
